Update eVTOL Aircraft Energy Consumption DRAFT.py
Formed the code as a function to allow for parametric study. Also made other small formatting changes to clean up the output section of the code. Did not introduce returns to the function, but I think that would be a great next step. For example, return Total Power and Total Energy to plot their relationship with flight distance and alpha.