Add files via upload
I have written a version of Matt and Hadi's code which I think should be kept separated. It allows the user to update given values if they would like. 
Additionally, it runs the main code inside a function which helps for parametric studies. I present an example at the bottom of this notebook. 
Let me know if you have any questions. Feel free to play around with it. Please reach out if you find bugs with its operation. 
Thanks,
Ethan